Refactored code and added to utils.py
